requirements.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.4
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
-e git+https://github.com/django-import-export/django-import-export.git@01ab0255d31ea63481ac3a78edd6fb9442e26320#egg=django_import_export
django-simple-history==3.0.0
django-suit @ https://github.com/darklow/django-suit/tarball/v2
djangorestframework==3.12.4
drf-yasg==1.20.0
et-xmlfile==1.1.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupPy==1.14
MarkupSafe==2.0.1
odfpy==1.4.1
openpyxl==3.0.7
packaging==21.0
Pillow==8.2.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
xlrd==2.0.1
xlwt==1.3.0
